requirements.txt
pyluach
lunarcalendar
hijridate
pandas
heniautos
requests
google-genai
openai
orjson
python-dateutil